requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
django-tastypie==0.14.2
djangorestframework==3.9.0
entrypoints==0.2.3
gunicorn==19.9.0
jsonschema==2.6.0
kiwisolver==1.0.1
notebook==5.7.0
numpy==1.15.3
pandas==0.23.4
plotly==3.3.0
protobuf==3.6.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
Pyrebase==3.0.27
python-dateutil==2.7.5
python-json-logger==0.1.8
python-jwt==2.0.1
python-mimeparse==1.6.0
pytz==2018.7
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.19.1
requests-toolbelt==0.7.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.8.2
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
tslearn==0.1.24
urllib3==1.24
virtualenv==16.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xlrd==1.1.0
